chore: size-gate splits for files grown past 2000 lines
lint's check_file_size.sh flagged lower_call/new.rs (2048) and
field_get_set/get_field_by_name_tail.rs (2006) after the #6530 additions.
Pure relocations per the split recipe:

- new.rs → new_ctor_args.rs: inline-ctor param binding/restore scope,
  the CaptureFill user-arg vs __perry_cap tail split (incl. the #6530
  new_site_args_carry_appended_caps check), rest/arguments packing,
  imported-ctor marshaling, and the standalone ctor-symbol call path.
  lower_call/mod.rs re-exports re-pointed; the dead zero-caller
  inline_constructor_param_values wrapper dropped.
- get_field_by_name_tail.rs → class_object_props.rs:
  instance_constructor_value — the 'constructor' special-key resolution
  arm (own field, WeakMap/WeakSet, #6530 class-object registry, vtable
  ctor method, boxed primitives, anon shapes, function-class table,
  ClassRef synthesis), returning None to fall through to the generic
  walk exactly as the inline arm did.
